+++ distri/geometric_distri.h
@@ -20,12 +20,13 @@
 #include "prob_distri.h"
 
 
-class geometric_Distri : public prob_Distri {
+class geometric_Distri : public prob_Distri
+{
     public:
         geometric_Distri(const qreal pro=0.5);  // constructor
-        qreal density( const int ) const;   // probability density function
-        qreal cumulat( const int ) const;   // cumulative probability function
-        qreal inverse( const qreal ) const;  // inverse cumulative prob. function
+        qreal density(const qreal) const;   // probability density function
+        qreal cumulat(const qreal) const;   // cumulative probability function
+        qreal inverse(const qreal) const;  // inverse cumulative prob. function
         int random( );                     // random prob. distributed variable
         int random(const qreal);            // random prob. distributed variable
     private:
